Refactor inset order optimization into its own class, InsetOrderOptimizer.
