+++ CMakeLists.txt
@@ -68,6 +68,7 @@ set(engine_SRCS # Except main.cpp.
     src/gcodeExport.cpp
     src/GCodePathConfig.cpp
     src/infill.cpp
+    src/InsetOrderOptimizer.cpp
     src/layerPart.cpp
     src/LayerPlan.cpp
     src/LayerPlanBuffer.cpp
